requirements.txt
streamlit
langchain
langchain_community
ollama
transformers
torch
pandas
PyMuPDF
python-docx
tqdm